feat: Chat 폴더 생성 & ChatInfo 분리 & 경로 수정
